run the event conversation test from both AzureStorage tests and ServiceBus tests
