+++ Test/DurableTask.AzureStorage.Tests/AzureStorageScenarioTests.cs
@@ -174,7 +174,7 @@ namespace DurableTask.AzureStorage.Tests
             {
                 await host.StartAsync();
 
-                var client = await host.StartOrchestrationAsync(typeof(Orchestrations.Greeter), "");
+                var client = await host.StartOrchestrationAsync(typeof(Test.Orchestrations.EventConversationOrchestration), "");
                 var status = await client.WaitForCompletionAsync(TimeSpan.FromSeconds(30));
 
                 Assert.AreEqual(OrchestrationStatus.Completed, status?.OrchestrationStatus);
@@ -2194,81 +2194,6 @@ namespace DurableTask.AzureStorage.Tests
                     await Task.WhenAll(tasks2);
 
                     return "OK";
-                }
-            }
-
-            [KnownType(typeof(Responder))]
-            internal class Greeter : TaskOrchestration<string, string>
-            {
-                private readonly TaskCompletionSource<string> tcs
-                    = new TaskCompletionSource<string>(TaskContinuationOptions.ExecuteSynchronously);
-
-                public async override Task<string> RunTask(OrchestrationContext context, string input)
-                {
-                    // start a responder orchestration
-                    var responderId = "responderId";
-                    var responderOrchestration = context.CreateSubOrchestrationInstance<string>(typeof(Responder), responderId, "Herkimer");
-
-                    // send the id of this orchestration to the responder
-                    var responderInstance = new OrchestrationInstance() { InstanceId = responderId };
-                    context.SendEvent(responderInstance, channelName, context.OrchestrationInstance.InstanceId);
-
-                    // wait for a response event 
-                    var message = await tcs.Task;
-                    Assert.AreEqual("hi from Herkimer", message);
-
-                    // tell the responder to stop listening, then wait for it to complete
-                    context.SendEvent(responderInstance, channelName, "stop");
-                    var receiverResult = await responderOrchestration;
-
-                    Assert.AreEqual("Herkimer is done", receiverResult);
-
-                    return "OK";
-                }
-
-                public override void OnEvent(OrchestrationContext context, string name, string input)
-                {
-                    if (name == channelName)
-                    {
-                        tcs.TrySetResult(input);
-                    }
-                }
-            }
-
-            private const string channelName = "conversation";
-
-            internal class Responder : TaskOrchestration<string, string>
-            {
-                private readonly TaskCompletionSource<string> tcs
-                    = new TaskCompletionSource<string>(TaskContinuationOptions.ExecuteSynchronously);
-
-                public async override Task<string> RunTask(OrchestrationContext context, string input)
-                {
-                    var message = await tcs.Task;
-
-                    if (message == "stop")
-                    {
-                        return $"{input} is done";
-                    }
-                    else
-                    {
-                        // send a message back to the sender
-                        var senderInstance = new OrchestrationInstance() { InstanceId = message };
-                        context.SendEvent(senderInstance, channelName, $"hi from {input}");
-
-                        // start over to wait for the next message
-                        context.ContinueAsNew(input);
-
-                        return "this value is meaningless";
-                    }
-                }
-
-                public override void OnEvent(OrchestrationContext context, string name, string input)
-                {
-                    if (name == channelName)
-                    {
-                        tcs.TrySetResult(input);
-                    }
                 }
             }
         }
