+++ test/DurableTask.Test.Orchestrations/SimpleOrchestrations.cs
@@ -14,6 +14,7 @@
 namespace DurableTask.Test.Orchestrations
 {
     using System;
+    using System.Runtime.Serialization;
     using System.Threading;
     using System.Threading.Tasks;
     using DurableTask.Core;
@@ -261,4 +262,87 @@ namespace DurableTask.Test.Orchestrations
             }
         }
     }
+
+    [KnownType(typeof(EventConversationOrchestration.Responder))]
+    public sealed class EventConversationOrchestration : TaskOrchestration<string, string>
+    {
+        private readonly TaskCompletionSource<string> tcs
+            = new TaskCompletionSource<string>(TaskContinuationOptions.ExecuteSynchronously);
+
+        // HACK: This is just a hack to communicate result of orchestration back to test
+        public static bool OkResult;
+
+        public async override Task<string> RunTask(OrchestrationContext context, string input)
+        {
+            // start a responder orchestration
+            var responderId = "responderId";
+            var responderOrchestration = context.CreateSubOrchestrationInstance<string>(typeof(Responder), responderId, "Herkimer");
+
+            // send the id of this orchestration to the responder
+            var responderInstance = new OrchestrationInstance() { InstanceId = responderId };
+            context.SendEvent(responderInstance, channelName, context.OrchestrationInstance.InstanceId);
+
+            // wait for a response event 
+            var message = await tcs.Task;
+            if (message != "hi from Herkimer")
+                throw new Exception("test failed");
+
+            // tell the responder to stop listening, then wait for it to complete
+            context.SendEvent(responderInstance, channelName, "stop");
+            var receiverResult = await responderOrchestration;
+
+            if (receiverResult != "Herkimer is done")
+                throw new Exception("test failed");
+
+            OkResult = true;
+
+            return "OK";
+        }
+
+        public override void OnEvent(OrchestrationContext context, string name, string input)
+        {
+            if (name == channelName)
+            {
+                tcs.TrySetResult(input);
+            }
+        }
+
+        private const string channelName = "conversation";
+
+        public class Responder : TaskOrchestration<string, string>
+        {
+            private readonly TaskCompletionSource<string> tcs
+                = new TaskCompletionSource<string>(TaskContinuationOptions.ExecuteSynchronously);
+
+            public async override Task<string> RunTask(OrchestrationContext context, string input)
+            {
+                var message = await tcs.Task;
+
+                if (message == "stop")
+                {
+                    return $"{input} is done";
+                }
+                else
+                {
+                    // send a message back to the sender
+                    var senderInstance = new OrchestrationInstance() { InstanceId = message };
+                    context.SendEvent(senderInstance, channelName, $"hi from {input}");
+
+                    // start over to wait for the next message
+                    context.ContinueAsNew(input);
+
+                    return "this value is meaningless";
+                }
+            }
+
+            public override void OnEvent(OrchestrationContext context, string name, string input)
+            {
+                if (name == channelName)
+                {
+                    tcs.TrySetResult(input);
+                }
+            }
+        }
+    }
+
 }
